fix: match every users.id FK column to the real (signed int) live schema
Same bug class as the product/invoice fix (57815c9), on users.id this
time: confirmed via SHOW CREATE TABLE on production, users.id is plain
SIGNED int(11), not Laravel's bigint id() default. All 13 migrations
with a FK to users.id already had a comment correctly stating this,
but every one of them still declared the column as
unsignedBigInteger() instead of integer() - the comment diagnosed the
problem but the code never actually implemented the fix. MySQL refuses
to constrain a mismatched-type FK (errno 150), which is what stopped
the live deploy at create_job_checklist_items_table right after the
prior product/invoice fix let it get further than before.

Fixed all 13 in one pass rather than one deploy attempt at a time:
jobs, job_photos, job_audit_logs, employees, attendances,
salary_payments, employee_documents, audit_logs,
job_checklist_items (the one currently blocking production),
vendor_bills, vendor_payments, daily_transactions,
merge_ticket_photos_into_job_photos. Also fixed create_users_table
itself (id() -> integer('id', true)) so a fresh install produces the
same schema production actually has, same as create_product_table /
create_invoice_table.

sessions.user_id (foreignId() with no ->constrained()) was left alone
- no FK constraint is created there, so no type-matching requirement
applies; not a crash risk, just an oversized column type.

Verified against a real local MySQL: full fresh migrate from zero
(all 71 migrations) plus the RolesAndPermissionsSeeder both pass
cleanly end to end. Full local suite still green (547 passed).

+++ database/migrations/2026_07_18_000007_create_jobs_table.php
@@ -14,18 +14,19 @@ return new class extends Migration
             $table->text('description')->nullable();
             $table->foreignId('machine_id')->nullable()->constrained('machines')->nullOnDelete();
             $table->string('site_name')->nullable();
-            // Plain integer() + explicit foreign(), not foreignId()->constrained(): this
-            // app's live `users.id` column predates Laravel's id() convention and is a
-            // signed `int`, not `unsigned bigint` - foreignId() would create a column
-            // type MySQL refuses as FK-incompatible with the real users table (only
-            // caught by testing against live MySQL; sqlite's weak typing let this pass
-            // silently in the automated test suite).
-            $table->unsignedBigInteger('created_by');
-            $table->unsignedBigInteger('assigned_to')->nullable();
+            // Plain integer() + explicit foreign(): this app's live `users.id`
+            // column predates Laravel's id() convention and is a signed
+            // `int` (confirmed via SHOW CREATE TABLE on production) - both
+            // foreignId() AND unsignedBigInteger() create a column type
+            // MySQL refuses as FK-incompatible with the real users table
+            // (only caught by testing against live MySQL; sqlite's weak
+            // typing let this pass silently in the automated test suite).
+            $table->integer('created_by');
+            $table->integer('assigned_to')->nullable();
             $table->enum('priority', ['low', 'medium', 'high', 'urgent'])->default('medium');
             $table->date('due_date')->nullable();
             $table->enum('status', ['pending_approval', 'assigned', 'in_progress', 'on_hold', 'completed', 'rejected']);
-            $table->unsignedBigInteger('decided_by')->nullable();
+            $table->integer('decided_by')->nullable();
             $table->timestamp('decided_at')->nullable();
             $table->text('rejection_reason')->nullable();
             $table->text('on_hold_reason')->nullable();
